eldritchmush/requirements.txt
evennia
django-allauth[socialaccount]>=0.58
